Refactor of packages to isolate subsystems
