+++ app/src/main/java/edu/uco/schambers4/octane/DataAccessObjects/Exercise/ExerciseRepository.java
@@ -1,8 +1,8 @@
-package edu.uco.schambers4.octane.DataAccessObjects;
+package edu.uco.schambers4.octane.DataAccessObjects.Exercise;
 
 import java.util.ArrayList;
-import edu.uco.schambers4.octane.Models.Exercise;
-import edu.uco.schambers4.octane.Models.Workout;
+import edu.uco.schambers4.octane.Models.Workout.Exercise;
+import edu.uco.schambers4.octane.Models.Workout.Workout;
 
 public interface ExerciseRepository {
