+++ app/src/main/java/edu/uco/schambers4/octane/DataAccessObjects/Exercise/MockExerciseRepository.java
@@ -1,10 +1,9 @@
-package edu.uco.schambers4.octane.DataAccessObjects;
+package edu.uco.schambers4.octane.DataAccessObjects.Exercise;
 
 import java.util.ArrayList;
 
-import edu.uco.schambers4.octane.DataAccessObjects.ExerciseRepository;
-import edu.uco.schambers4.octane.Models.Exercise;
-import edu.uco.schambers4.octane.Models.Workout;
+import edu.uco.schambers4.octane.Models.Workout.Exercise;
+import edu.uco.schambers4.octane.Models.Workout.Workout;
 
 /**
  * Created by gordon on 10/26/15.
