+++ app/src/main/java/edu/uco/schambers4/octane/Fragments/Workouts/WorkoutListFragment.java
@@ -1,16 +1,14 @@
 package edu.uco.schambers4.octane.Fragments.Workouts;
 
 import android.app.Fragment;
-import android.app.FragmentTransaction;
 import android.app.ListFragment;
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ListView;
 
 import edu.uco.schambers4.octane.Activities.MainActivity;
-import edu.uco.schambers4.octane.Models.WorkoutAdapter;
-import edu.uco.schambers4.octane.Models.WorkoutContainer;
+import edu.uco.schambers4.octane.Models.Workout.WorkoutAdapter;
+import edu.uco.schambers4.octane.Models.Workout.WorkoutContainer;
 import edu.uco.schambers4.octane.R;
 
 public class WorkoutListFragment extends ListFragment {
