+++ app/src/main/java/edu/uco/schambers4/octane/Models/Workout/ExerciseAdapter.java
@@ -1,4 +1,4 @@
-package edu.uco.schambers4.octane.Models;
+package edu.uco.schambers4.octane.Models.Workout;
 
 import android.content.Context;
 import android.view.LayoutInflater;
