+++ app/src/main/java/edu/uco/schambers4/octane/Models/Workout/ExerciseMeasure.java
@@ -1,4 +1,4 @@
-package edu.uco.schambers4.octane.Models;
+package edu.uco.schambers4.octane.Models.Workout;
 
 public class ExerciseMeasure {
