--- app/src/main/res/layout-land/activity_exercises.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:tools="http://schemas.android.com/tools"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              android:orientation="horizontal">
-
-    <fragment
-        android:id="@+id/list"
-        android:name="edu.uco.schambers4.octane.Fragments.Exercises.ExerciseDetailFragment"
-        android:layout_width="0dp"
-        android:layout_height="match_parent"
-        android:layout_weight="1"
-        tools:layout="@layout/fragment_exercise_detail"/>
-
-
-    <FrameLayout
-        android:id="@+id/exercise_item_details"
-        android:layout_width="0px"
-        android:layout_height="match_parent"
-        android:layout_weight="2"
-        android:background="?android:attr/detailsElementBackground"
-        />
-
-</LinearLayout>
